feat(am.6): rebase thin-socket onto integrate/phase-AM
Rebases all AM.6 changes (thin socket GH monitor boundary, routing
module extraction, CI monitor restructure) onto the integrate/phase-AM
base that includes AM.5 (health state split).

Conflict resolutions:
- socket.rs: removed AM.5 test-only import stubs (superseded by AM.6
  thin boundary commit which relocated them to gh_monitor_router_tests.rs)
- ci_monitor/mod.rs: kept gh_alerts module name from AM.5, added new
  polling/routing modules from AM.6, applied Unix gating from AM.6
- ci_monitor/routing.rs: replaced circular health delegation with
  direct implementation (from AM.5 health.rs); updated alerts references
  to use gh_alerts module name
- ci_monitor/gh_monitor.rs, polling.rs: updated super::alerts imports
  to super::gh_alerts to match AM.5 module rename
