+++ crates/atm-daemon/src/plugins/ci_monitor/test_support.rs
@@ -1,5 +1,6 @@
 use std::path::Path;
 
+use agent_team_mail_core::schema::InboxMessage;
 use tempfile::TempDir;
 
 pub(crate) fn write_gh_monitor_config(home: &Path, team: &str) {
@@ -83,6 +84,64 @@ impl Drop for EnvGuard {
             }
         }
     }
+}
+
+pub(crate) fn write_hook_auth_team_config(
+    home_dir: &Path,
+    team: &str,
+    lead: &str,
+    members: &[&str],
+) {
+    let team_dir = home_dir.join(".claude/teams").join(team);
+    std::fs::create_dir_all(&team_dir).unwrap();
+    let mut member_values = Vec::new();
+    for m in members {
+        member_values.push(serde_json::json!({
+            "agentId": format!("{m}@{team}"),
+            "name": m,
+            "agentType": "general-purpose",
+            "model": "unknown",
+            "joinedAt": 1739284800000u64,
+            "cwd": home_dir.to_string_lossy().to_string(),
+            "subscriptions": []
+        }));
+    }
+    let config = serde_json::json!({
+        "name": team,
+        "description": "test team",
+        "createdAt": 1739284800000u64,
+        "leadAgentId": format!("{lead}@{team}"),
+        "leadSessionId": "test-lead-session",
+        "members": member_values,
+    });
+    {
+        use std::io::Write;
+        let config_path = team_dir.join("config.json");
+        let config_bytes = serde_json::to_string_pretty(&config).unwrap();
+        let file = std::fs::File::create(&config_path).unwrap();
+        let mut writer = std::io::BufWriter::new(&file);
+        writer.write_all(config_bytes.as_bytes()).unwrap();
+        writer.flush().unwrap();
+        file.sync_all().unwrap();
+    }
+}
+
+#[cfg(unix)]
+pub(crate) fn read_team_inbox_messages(
+    home_dir: &Path,
+    team: &str,
+    agent: &str,
+) -> Vec<InboxMessage> {
+    let path = home_dir
+        .join(".claude/teams")
+        .join(team)
+        .join("inboxes")
+        .join(format!("{agent}.json"));
+    if !path.exists() {
+        return Vec::new();
+    }
+    serde_json::from_str::<Vec<InboxMessage>>(&std::fs::read_to_string(path).unwrap())
+        .unwrap_or_default()
 }
 
 #[cfg(unix)]
